refactored elektronLogging to logging
