+++ def/taskUtils.go
@@ -26,8 +26,8 @@ import (
 	"github.com/mash/gokmeans"
 	"github.com/montanaflynn/stats"
 	log "github.com/sirupsen/logrus"
-	elekLog "github.com/spdfg/elektron/elektronLogging"
-	elekLogTypes "github.com/spdfg/elektron/elektronLogging/types"
+	elekLog "github.com/spdfg/elektron/logging"
+	elekLogTypes "github.com/spdfg/elektron/logging/types"
 )
 
 // Information about a cluster of tasks.
