+++ logging/schedPolicySwitchLogger.go
@@ -1,4 +1,4 @@
-package elektronLogging
+package logging
 
 import (
 	"os"
@@ -8,13 +8,13 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-type SchedWindowLogger struct {
+type SchedPolicySwitchLogger struct {
 	loggerImpl
 }
 
-func NewSchedWindowLogger(b *baseLogData, logType int, prefix string,
-	logger *log.Logger, logDir *logDirectory) *SchedWindowLogger {
-	sLog := &SchedWindowLogger{}
+func NewSchedPolicySwitchLogger(b *baseLogData, logType int, prefix string,
+	logger *log.Logger, logDir *logDirectory) *SchedPolicySwitchLogger {
+	sLog := &SchedPolicySwitchLogger{}
 	sLog.logType = logType
 	sLog.logDir = logDir
 	sLog.next = nil
@@ -24,9 +24,9 @@ func NewSchedWindowLogger(b *baseLogData, logType int, prefix string,
 	return sLog
 }
 
-func (sLog SchedWindowLogger) Log(logType int, level log.Level, message string) {
+func (sLog SchedPolicySwitchLogger) Log(logType int, level log.Level, message string) {
 	if sLog.logType == logType {
-		if config.SchedWindowConfig.Enabled {
+		if config.SPSConfig.Enabled {
 			if sLog.allowOnConsole {
 				sLog.logger.SetOutput(os.Stdout)
 				sLog.logger.WithFields(sLog.data).Log(level, message)
@@ -36,7 +36,6 @@ func (sLog SchedWindowLogger) Log(logType int, level log.Level, message string) 
 			sLog.logger.WithFields(sLog.data).Log(level, message)
 		}
 	}
-	// Forwarding to next logger
 	if sLog.next != nil {
 		sLog.next.Log(logType, level, message)
 	} else {
@@ -45,9 +44,9 @@ func (sLog SchedWindowLogger) Log(logType int, level log.Level, message string) 
 	}
 }
 
-func (sLog SchedWindowLogger) Logf(logType int, level log.Level, msgFmtString string, args ...interface{}) {
+func (sLog SchedPolicySwitchLogger) Logf(logType int, level log.Level, msgFmtString string, args ...interface{}) {
 	if sLog.logType == logType {
-		if config.SchedWindowConfig.Enabled {
+		if config.SPSConfig.Enabled {
 			if sLog.allowOnConsole {
 				sLog.logger.SetOutput(os.Stdout)
 				sLog.logger.WithFields(sLog.data).Logf(level, msgFmtString, args...)
@@ -57,6 +56,7 @@ func (sLog SchedWindowLogger) Logf(logType int, level log.Level, msgFmtString st
 			sLog.logger.WithFields(sLog.data).Logf(level, msgFmtString, args...)
 		}
 	}
+	// Forwarding to next logger
 	if sLog.next != nil {
 		sLog.next.Logf(logType, level, msgFmtString, args...)
 	} else {
@@ -65,16 +65,16 @@ func (sLog SchedWindowLogger) Logf(logType int, level log.Level, msgFmtString st
 	}
 }
 
-func (sLog *SchedWindowLogger) createLogFile(prefix string) {
-	if config.SchedWindowConfig.Enabled {
-		filename := strings.Join([]string{prefix, config.SchedWindowConfig.FilenameExtension}, "")
+func (sLog *SchedPolicySwitchLogger) createLogFile(prefix string) {
+	if config.SPSConfig.Enabled {
+		filename := strings.Join([]string{prefix, config.SPSConfig.FilenameExtension}, "")
 		dirName := sLog.logDir.getDirName()
 		if dirName != "" {
 			if logFile, err := os.Create(filepath.Join(dirName, filename)); err != nil {
 				log.Fatal("Unable to create logFile: ", err)
 			} else {
 				sLog.logFile = logFile
-				sLog.allowOnConsole = config.SchedWindowConfig.AllowOnConsole
+				sLog.allowOnConsole = config.SPSConfig.AllowOnConsole
 			}
 		}
 	}
