+++ logging/schedTraceLogger.go
@@ -1,4 +1,4 @@
-package elektronLogging
+package logging
 
 import (
 	"os"
@@ -8,13 +8,13 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-type SchedPolicySwitchLogger struct {
+type SchedTraceLogger struct {
 	loggerImpl
 }
 
-func NewSchedPolicySwitchLogger(b *baseLogData, logType int, prefix string,
-	logger *log.Logger, logDir *logDirectory) *SchedPolicySwitchLogger {
-	sLog := &SchedPolicySwitchLogger{}
+func NewSchedTraceLogger(b *baseLogData, logType int, prefix string,
+	logger *log.Logger, logDir *logDirectory) *SchedTraceLogger {
+	sLog := &SchedTraceLogger{}
 	sLog.logType = logType
 	sLog.logDir = logDir
 	sLog.next = nil
@@ -24,9 +24,9 @@ func NewSchedPolicySwitchLogger(b *baseLogData, logType int, prefix string,
 	return sLog
 }
 
-func (sLog SchedPolicySwitchLogger) Log(logType int, level log.Level, message string) {
+func (sLog SchedTraceLogger) Log(logType int, level log.Level, message string) {
 	if sLog.logType == logType {
-		if config.SPSConfig.Enabled {
+		if config.SchedTraceConfig.Enabled {
 			if sLog.allowOnConsole {
 				sLog.logger.SetOutput(os.Stdout)
 				sLog.logger.WithFields(sLog.data).Log(level, message)
@@ -44,9 +44,9 @@ func (sLog SchedPolicySwitchLogger) Log(logType int, level log.Level, message st
 	}
 }
 
-func (sLog SchedPolicySwitchLogger) Logf(logType int, level log.Level, msgFmtString string, args ...interface{}) {
+func (sLog SchedTraceLogger) Logf(logType int, level log.Level, msgFmtString string, args ...interface{}) {
 	if sLog.logType == logType {
-		if config.SPSConfig.Enabled {
+		if config.SchedTraceConfig.Enabled {
 			if sLog.allowOnConsole {
 				sLog.logger.SetOutput(os.Stdout)
 				sLog.logger.WithFields(sLog.data).Logf(level, msgFmtString, args...)
@@ -65,16 +65,16 @@ func (sLog SchedPolicySwitchLogger) Logf(logType int, level log.Level, msgFmtStr
 	}
 }
 
-func (sLog *SchedPolicySwitchLogger) createLogFile(prefix string) {
-	if config.SPSConfig.Enabled {
-		filename := strings.Join([]string{prefix, config.SPSConfig.FilenameExtension}, "")
+func (sLog *SchedTraceLogger) createLogFile(prefix string) {
+	if config.SchedTraceConfig.Enabled {
+		filename := strings.Join([]string{prefix, config.SchedTraceConfig.FilenameExtension}, "")
 		dirName := sLog.logDir.getDirName()
 		if dirName != "" {
 			if logFile, err := os.Create(filepath.Join(dirName, filename)); err != nil {
 				log.Fatal("Unable to create logFile: ", err)
 			} else {
 				sLog.logFile = logFile
-				sLog.allowOnConsole = config.SPSConfig.AllowOnConsole
+				sLog.allowOnConsole = config.SchedTraceConfig.AllowOnConsole
 			}
 		}
 	}
