+++ logging/schedWindowLogger.go
@@ -1,4 +1,4 @@
-package elektronLogging
+package logging
 
 import (
 	"os"
@@ -8,13 +8,13 @@ import (
 	log "github.com/sirupsen/logrus"
 )
 
-type SchedTraceLogger struct {
+type SchedWindowLogger struct {
 	loggerImpl
 }
 
-func NewSchedTraceLogger(b *baseLogData, logType int, prefix string,
-	logger *log.Logger, logDir *logDirectory) *SchedTraceLogger {
-	sLog := &SchedTraceLogger{}
+func NewSchedWindowLogger(b *baseLogData, logType int, prefix string,
+	logger *log.Logger, logDir *logDirectory) *SchedWindowLogger {
+	sLog := &SchedWindowLogger{}
 	sLog.logType = logType
 	sLog.logDir = logDir
 	sLog.next = nil
@@ -24,9 +24,9 @@ func NewSchedTraceLogger(b *baseLogData, logType int, prefix string,
 	return sLog
 }
 
-func (sLog SchedTraceLogger) Log(logType int, level log.Level, message string) {
+func (sLog SchedWindowLogger) Log(logType int, level log.Level, message string) {
 	if sLog.logType == logType {
-		if config.SchedTraceConfig.Enabled {
+		if config.SchedWindowConfig.Enabled {
 			if sLog.allowOnConsole {
 				sLog.logger.SetOutput(os.Stdout)
 				sLog.logger.WithFields(sLog.data).Log(level, message)
@@ -36,6 +36,7 @@ func (sLog SchedTraceLogger) Log(logType int, level log.Level, message string) {
 			sLog.logger.WithFields(sLog.data).Log(level, message)
 		}
 	}
+	// Forwarding to next logger
 	if sLog.next != nil {
 		sLog.next.Log(logType, level, message)
 	} else {
@@ -44,9 +45,9 @@ func (sLog SchedTraceLogger) Log(logType int, level log.Level, message string) {
 	}
 }
 
-func (sLog SchedTraceLogger) Logf(logType int, level log.Level, msgFmtString string, args ...interface{}) {
+func (sLog SchedWindowLogger) Logf(logType int, level log.Level, msgFmtString string, args ...interface{}) {
 	if sLog.logType == logType {
-		if config.SchedTraceConfig.Enabled {
+		if config.SchedWindowConfig.Enabled {
 			if sLog.allowOnConsole {
 				sLog.logger.SetOutput(os.Stdout)
 				sLog.logger.WithFields(sLog.data).Logf(level, msgFmtString, args...)
@@ -56,7 +57,6 @@ func (sLog SchedTraceLogger) Logf(logType int, level log.Level, msgFmtString str
 			sLog.logger.WithFields(sLog.data).Logf(level, msgFmtString, args...)
 		}
 	}
-	// Forwarding to next logger
 	if sLog.next != nil {
 		sLog.next.Logf(logType, level, msgFmtString, args...)
 	} else {
@@ -65,16 +65,16 @@ func (sLog SchedTraceLogger) Logf(logType int, level log.Level, msgFmtString str
 	}
 }
 
-func (sLog *SchedTraceLogger) createLogFile(prefix string) {
-	if config.SchedTraceConfig.Enabled {
-		filename := strings.Join([]string{prefix, config.SchedTraceConfig.FilenameExtension}, "")
+func (sLog *SchedWindowLogger) createLogFile(prefix string) {
+	if config.SchedWindowConfig.Enabled {
+		filename := strings.Join([]string{prefix, config.SchedWindowConfig.FilenameExtension}, "")
 		dirName := sLog.logDir.getDirName()
 		if dirName != "" {
 			if logFile, err := os.Create(filepath.Join(dirName, filename)); err != nil {
 				log.Fatal("Unable to create logFile: ", err)
 			} else {
 				sLog.logFile = logFile
-				sLog.allowOnConsole = config.SchedTraceConfig.AllowOnConsole
+				sLog.allowOnConsole = config.SchedWindowConfig.AllowOnConsole
 			}
 		}
 	}
