+++ pcp/pcp.go
@@ -25,8 +25,8 @@ import (
 	"time"
 
 	log "github.com/sirupsen/logrus"
-	elekLog "github.com/spdfg/elektron/elektronLogging"
-	elekLogTypes "github.com/spdfg/elektron/elektronLogging/types"
+	elekLog "github.com/spdfg/elektron/logging"
+	elekLogTypes "github.com/spdfg/elektron/logging/types"
 )
 
 func Start(quit chan struct{}, logging *bool, pcpConfigFile string) {
