+++ powerCap/progressiveExtrema.go
@@ -32,8 +32,8 @@ import (
 
 	log "github.com/sirupsen/logrus"
 	"github.com/spdfg/elektron/constants"
-	elekLog "github.com/spdfg/elektron/elektronLogging"
-	elekLogTypes "github.com/spdfg/elektron/elektronLogging/types"
+	elekLog "github.com/spdfg/elektron/logging"
+	elekLogTypes "github.com/spdfg/elektron/logging/types"
 	"github.com/spdfg/elektron/pcp"
 	"github.com/spdfg/elektron/rapl"
 	"github.com/spdfg/elektron/utilities"
