+++ schedulers/schedPolicy.go
@@ -26,8 +26,8 @@ import (
 	sched "github.com/mesos/mesos-go/api/v0/scheduler"
 	log "github.com/sirupsen/logrus"
 	"github.com/spdfg/elektron/def"
-	elekLog "github.com/spdfg/elektron/elektronLogging"
-	elekLogTypes "github.com/spdfg/elektron/elektronLogging/types"
+	elekLog "github.com/spdfg/elektron/logging"
+	elekLogTypes "github.com/spdfg/elektron/logging/types"
 )
 
 type SchedPolicyContext interface {
